+++ packages/cli/src/core/types/index.d.ts
@@ -1,2 +1,0 @@
-declare module "sync-disk-cache";
-declare module "solc";