+++ packages/cli/src/lib/types.ts
@@ -18,3 +18,25 @@ export type IProcessAppTypeArgs = {
   hasEthers: boolean;
   otherDeps: Set<string>;
 }
+
+export type ApiKey = {
+  id: string;
+  key: string;
+  walletAddresses: string[];
+  domains: string[];
+  services?: [
+    {
+      name: string;
+      targetAddresses: string[];
+    },
+  ];
+};
+
+export type ApiResponse = {
+  data: ApiKey | null;
+  error: {
+    code: string;
+    statusCode: number;
+    message: string;
+  };
+};